+++ lua/plugins/diffview.lua
@@ -1,0 +1,10 @@
+-- lua/plugins/diffview.lua
+
+return {
+    'sindrets/diffview.nvim',
+    dependencies = { 'nvim-lua/plenary.nvim' },
+    cmd    = { 'DiffviewOpen', 'DiffviewClose', 'DiffviewFileHistory' },
+    config = function()
+        require('diffview').setup()
+    end,
+}