+++ lua/plugins/todo-comments.lua
@@ -1,0 +1,10 @@
+-- lua/plugins/todo-comments.lua
+
+return {
+    'folke/todo-comments.nvim',
+    dependencies = { 'nvim-lua/plenary.nvim' },
+    event  = { 'BufReadPre', 'BufNewFile' },
+    config = function()
+        require('todo-comments').setup()
+    end,
+}